# Implement object-based token_get_all() alternative