--- a/ext/tokenizer/tokenizer.c
+++ b/ext/tokenizer/tokenizer.c
@@ -35,10 +35,14 @@
 #define zendcursor LANG_SCNG(yy_cursor)
 #define zendlimit  LANG_SCNG(yy_limit)
 
-#define TOKEN_PARSE 				1
+#define TOKEN_PARSE     (1 << 0)
+#define TOKEN_AS_OBJECT (1 << 1)
+
+zend_class_entry *php_token_ce;
 
 void tokenizer_token_get_all_register_constants(INIT_FUNC_ARGS) {
 	REGISTER_LONG_CONSTANT("TOKEN_PARSE", TOKEN_PARSE, CONST_CS|CONST_PERSISTENT);
+	REGISTER_LONG_CONSTANT("TOKEN_AS_OBJECT", TOKEN_AS_OBJECT, CONST_CS|CONST_PERSISTENT);
 }
 
 /* {{{ tokenizer_functions[]
@@ -76,8 +80,37 @@ ZEND_GET_MODULE(tokenizer)
  */
 PHP_MINIT_FUNCTION(tokenizer)
 {
+	zend_class_entry ce;
+	zend_string *name;
+	zval default_val;
+	ZVAL_UNDEF(&default_val);
+
 	tokenizer_register_constants(INIT_FUNC_ARGS_PASSTHRU);
 	tokenizer_token_get_all_register_constants(INIT_FUNC_ARGS_PASSTHRU);
+
+	INIT_CLASS_ENTRY(ce, "PhpToken", NULL);
+	php_token_ce = zend_register_internal_class(&ce);
+
+	name = zend_string_init("id", sizeof("id") - 1, 1);
+	zend_declare_typed_property(php_token_ce, name, &default_val, ZEND_ACC_PUBLIC, NULL,
+		(zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
+	zend_string_release(name);
+
+	name = zend_string_init("text", sizeof("text") - 1, 1);
+	zend_declare_typed_property(php_token_ce, name, &default_val, ZEND_ACC_PUBLIC, NULL,
+		(zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_STRING));
+	zend_string_release(name);
+
+	name = zend_string_init("line", sizeof("line") - 1, 1);
+	zend_declare_typed_property(php_token_ce, name, &default_val, ZEND_ACC_PUBLIC, NULL,
+		(zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
+	zend_string_release(name);
+
+	name = zend_string_init("pos", sizeof("pos") - 1, 1);
+	zend_declare_typed_property(php_token_ce, name, &default_val, ZEND_ACC_PUBLIC, NULL,
+		(zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
+	zend_string_release(name);
+
 	return SUCCESS;
 }
 /* }}} */
@@ -92,29 +125,36 @@ PHP_MINFO_FUNCTION(tokenizer)
 }
 /* }}} */
 
-static void add_token(zval *return_value, int token_type,
-		unsigned char *text, size_t leng, int lineno) {
-	if (token_type >= 256) {
-		zval keyword;
-		array_init(&keyword);
-		add_next_index_long(&keyword, token_type);
-		if (leng == 1) {
-			add_next_index_str(&keyword, ZSTR_CHAR(text[0]));
-		} else {
-			add_next_index_stringl(&keyword, (char *) text, leng);
-		}
-		add_next_index_long(&keyword, lineno);
-		add_next_index_zval(return_value, &keyword);
+static inline zend_string *make_str(unsigned char *text, size_t leng) {
+	if (leng == 1) {
+		return ZSTR_CHAR(text[0]);
 	} else {
-		if (leng == 1) {
-			add_next_index_str(return_value, ZSTR_CHAR(text[0]));
-		} else {
-			add_next_index_stringl(return_value, (char *) text, leng);
-		}
+		return zend_string_init((char *) text, leng, 0);
 	}
 }
 
-static zend_bool tokenize(zval *return_value, zend_string *source)
+static void add_token(zval *return_value, int token_type,
+		unsigned char *text, size_t leng, int lineno, zend_bool as_object) {
+	zval token;
+	if (as_object) {
+		zend_object *obj = zend_objects_new(php_token_ce);
+		ZVAL_OBJ(&token, obj);
+		ZVAL_LONG(OBJ_PROP_NUM(obj, 0), token_type);
+		ZVAL_STR(OBJ_PROP_NUM(obj, 1), make_str(text, leng));
+		ZVAL_LONG(OBJ_PROP_NUM(obj, 2), lineno);
+		ZVAL_LONG(OBJ_PROP_NUM(obj, 3), text - LANG_SCNG(yy_start));
+	} else if (token_type >= 256) {
+		array_init(&token);
+		add_next_index_long(&token, token_type);
+		add_next_index_str(&token, make_str(text, leng));
+		add_next_index_long(&token, lineno);
+	} else {
+		ZVAL_STR(&token, make_str(text, leng));
+	}
+	zend_hash_next_index_insert_new(Z_ARRVAL_P(return_value), &token);
+}
+
+static zend_bool tokenize(zval *return_value, zend_string *source, zend_bool as_object)
 {
 	zval source_zval;
 	zend_lex_state original_lex_state;
@@ -135,7 +175,7 @@ static zend_bool tokenize(zval *return_value, zend_string *source)
 	array_init(return_value);
 
 	while ((token_type = lex_scan(&token, NULL))) {
-		add_token(return_value, token_type, zendtext, zendleng, token_line);
+		add_token(return_value, token_type, zendtext, zendleng, token_line, as_object);
 
 		if (Z_TYPE(token) != IS_UNDEF) {
 			zval_ptr_dtor_nogc(&token);
@@ -151,7 +191,7 @@ static zend_bool tokenize(zval *return_value, zend_string *source)
 				/* fetch the rest into a T_INLINE_HTML */
 				if (zendcursor != zendlimit) {
 					add_token(return_value, T_INLINE_HTML,
-						zendcursor, zendlimit - zendcursor, token_line);
+						zendcursor, zendlimit - zendcursor, token_line, as_object);
 				}
 				break;
 			}
@@ -173,42 +213,49 @@ static zend_bool tokenize(zval *return_value, zend_string *source)
 	return 1;
 }
 
+struct event_context {
+	zval *tokens;
+	zend_bool as_object;
+};
+
 void on_event(zend_php_scanner_event event, int token, int line, void *context)
 {
-	zval *token_stream = (zval *) context;
+	struct event_context *ctx = context;
 	HashTable *tokens_ht;
 	zval *token_zv;
 
 	switch (event) {
 		case ON_TOKEN:
-			{
-				if (token == END) break;
-				/* Special cases */
-				if (token == ';' && LANG_SCNG(yy_leng) > 1) { /* ?> or ?>\n or ?>\r\n */
-					token = T_CLOSE_TAG;
-				} else if (token == T_ECHO && LANG_SCNG(yy_leng) == sizeof("<?=") - 1) {
-					token = T_OPEN_TAG_WITH_ECHO;
-				}
-				add_token(token_stream, token, LANG_SCNG(yy_text), LANG_SCNG(yy_leng), line);
+			if (token == END) break;
+			/* Special cases */
+			if (token == ';' && LANG_SCNG(yy_leng) > 1) { /* ?> or ?>\n or ?>\r\n */
+				token = T_CLOSE_TAG;
+			} else if (token == T_ECHO && LANG_SCNG(yy_leng) == sizeof("<?=") - 1) {
+				token = T_OPEN_TAG_WITH_ECHO;
 			}
+			add_token(ctx->tokens, token,
+				LANG_SCNG(yy_text), LANG_SCNG(yy_leng), line, ctx->as_object);
 			break;
 		case ON_FEEDBACK:
-			tokens_ht = Z_ARRVAL_P(token_stream);
+			tokens_ht = Z_ARRVAL_P(ctx->tokens);
 			token_zv = zend_hash_index_find(tokens_ht, zend_hash_num_elements(tokens_ht) - 1);
-			if (token_zv && Z_TYPE_P(token_zv) == IS_ARRAY) {
+			ZEND_ASSERT(token_zv);
+			if (Z_TYPE_P(token_zv) == IS_ARRAY) {
 				ZVAL_LONG(zend_hash_index_find(Z_ARRVAL_P(token_zv), 0), token);
+			} else {
+				zend_update_property_long(php_token_ce, token_zv, "type", sizeof("type")-1, token);
 			}
 			break;
 		case ON_STOP:
 			if (LANG_SCNG(yy_cursor) != LANG_SCNG(yy_limit)) {
-				add_token(token_stream, T_INLINE_HTML, LANG_SCNG(yy_cursor),
-					LANG_SCNG(yy_limit) - LANG_SCNG(yy_cursor), CG(zend_lineno));
+				add_token(ctx->tokens, T_INLINE_HTML, LANG_SCNG(yy_cursor),
+					LANG_SCNG(yy_limit) - LANG_SCNG(yy_cursor), CG(zend_lineno), ctx->as_object);
 			}
 			break;
 	}
 }
 
-static zend_bool tokenize_parse(zval *return_value, zend_string *source)
+static zend_bool tokenize_parse(zval *return_value, zend_string *source, zend_bool as_object)
 {
 	zval source_zval;
 	zend_lex_state original_lex_state;
@@ -222,14 +269,18 @@ static zend_bool tokenize_parse(zval *return_value, zend_string *source)
 	zend_save_lexical_state(&original_lex_state);
 
 	if ((success = (zend_prepare_string_for_scanning(&source_zval, "") == SUCCESS))) {
+		struct event_context ctx;
 		zval token_stream;
 		array_init(&token_stream);
+
+		ctx.tokens = &token_stream;
+		ctx.as_object = as_object;
 
 		CG(ast) = NULL;
 		CG(ast_arena) = zend_arena_create(1024 * 32);
 		LANG_SCNG(yy_state) = yycINITIAL;
 		LANG_SCNG(on_event) = on_event;
-		LANG_SCNG(on_event_context) = &token_stream;
+		LANG_SCNG(on_event_context) = &ctx;
 
 		if((success = (zendparse() == SUCCESS))) {
 			ZVAL_COPY_VALUE(return_value, &token_stream);
@@ -267,9 +318,9 @@ PHP_FUNCTION(token_get_all)
 	ZEND_PARSE_PARAMETERS_END();
 
 	if (flags & TOKEN_PARSE) {
-		success = tokenize_parse(return_value, source);
+		success = tokenize_parse(return_value, source, (flags & TOKEN_AS_OBJECT) != 0);
 	} else {
-		success = tokenize(return_value, source);
+		success = tokenize(return_value, source, (flags & TOKEN_AS_OBJECT) != 0);
 		/* Normal token_get_all() should not throw. */
 		zend_clear_exception();
 	}