--- a/ext/tokenizer/tokenizer_data_gen.sh
+++ b/ext/tokenizer/tokenizer_data_gen.sh
@@ -71,7 +71,7 @@ awk '
 
 echo '
 	}
-	return "UNKNOWN";
+	return NULL;
 }
 ' >> $outfile