@nikic 2020-02-12
Pushed another commit that adds local string interning -- may or may not want to enable that, but this is just a technical detail.

Some results for token_get_all() on a large PHP file (100x):

```
Without interning: 
    Default:
        Time: 0.42794013023376s
        Memory Usage: 14.037727355957MB
    TOKEN_AS_OBJECT:
        Time: 0.32422184944153s
        Memory Usage: 8.0559387207031MB
With interning: 
    Default:
        Time: 0.41672205924988s
        Memory Usage: 13.329658508301MB
    TOKEN_AS_OBJECT:
        Time: 0.35216188430786s
        Memory Usage: 7.3478698730469MB
```

We can see that TOKEN_AS_OBJECT drops memory usage by ~40% and improves performance by ~20%.

Interning further drops memory usage by 8% and regresses performance by 9%. As token_get_all() is usually "fast enough" (not the bottleneck at all), the memory usage saving is probably more valuable.