@staabm 2020-02-12
Should the class be different namend then PhpToken to reduce potential of conflicting with already existing userland classes?