@nikic 2020-02-12
@staabm No, I don't care about that at all. But if there is a better name (on it's own merits, not because of concerns about conflicts), then of course we can consider that -- I'm not particularly attached to this one.