@staabm 2020-02-12
I see. I would suggest `PhpParserToken` because „token“ itself is a term which is used in a lot of different contexts