@marcioAlmada 2020-02-13
Hi 😄 

This is indeed already much better than the `array|string` token mess. But the current interface still has a problem that always annoyed me, which is to decide how to compare the tokens:

```php
// before:

if (is_string($token) && $token === ';'); // === 59 would be worst here
else (is_array(($token) && $token[0] === T_OBJECT_OPERATOR);

// now

if ($token->text === ';');
else ($token->id === T_OBJECT_OPERATOR);
```

Would it be damaging to have a single field for both id and text? Or to have an method `is(int|string $id)` method to handle the comparison?

```php
if ($token->is(';'));
else ($token->is(T_OBJECT_OPERATOR));
```

Anyway it will be easy to wrap the token instances into a class with new methods on userland, at the cost of having two objects per token instead of one.