@nikic 2020-02-13
@marcioAlmada I don't see how a single field for id/text would work. For tokens with more than one character, those do not have a one-to-one relationship.

I think it's fine to compare id/text depending on whether it's a single-char token or a mutli-char token. An alternative would be to declare your own `define('T_SEMICOLON', \ord(';'))`, if you would like to base everything on `T_*` constants.

Having an `is()` method still may be useful -- I think not so much for consolidating the check for id/text, but to support the common extension of specifying multiple allowed token types:

```
public function is(int|string|array $type) {
    if (is_array($type)) {
        foreach ($type as $singleType) {
            if ($this->is($singleType)) {
                return true;
            }
        }
        return false;
    } else if (is_string($type)) {
        return $this->text === $type;
    } else {
        return $this->id === $type,
    }
}
```

This could be more efficient internally. Another potentially useful method is `isIgnorable()`, which would basically be the same as `is([T_WHITESPACE, T_COMMENT, T_DOC_COMMENT, T_OPEN_TAG, T_BAD_CHARACTER])`, just more efficient and matching PHP semantics. Another one is `getTokenName()`, so you don't have to go through `token_name()` for debug output.

However, I'm not entirely sure it's a good idea to include these initially. The previous discussion on this topic (https://externals.io/message/98610) degenerated because someone decided that as soon as an "object" is somehow involved, we need some ridiculously over-engineered object-oriented interface instead. I would like to avoid this.