@nikic 2020-02-13
RFC: https://wiki.php.net/rfc/token_as_object

I've included additional methods as an open question.