@nicolas-grekas 2020-02-13
Nice :)
Would it make sense to make the new behavior polyfillable, i.e. use a new function (or static method) instead of a flag?