@nikic 2020-02-13
@nicolas-grekas Probably makes sense... I've added it as an open question for now: https://wiki.php.net/rfc/token_as_object#construction_method

I was originally concerned about "combinatorial explosion" if each new flag gets a new function (or rather, multiplies all existing functions by two), but we don't really add new flags to this functionality that often...