@marcioAlmada 2020-02-13
> RFC: https://wiki.php.net/rfc/token_as_object
> 
> I've included additional methods as an open question.

Agreed. These methods seem enough as a baseline for most people doing tokenizing/parsing work without compromising too much, so +1

Maybe the `isIgnorable` method could be further optimized. It's very common to see this method called *at least* once per token on hot paths, if the parsing method backtracks this can be much more, perhaps.

Probably my last suggestion on this, as the RFC and patch already looks good IMMO. A new public property to compute the ignorable state on construction, could be doable:

```diff
class PhpToken {
    /** One of the T_* constants, or an integer < 256 representing a single-char token. */
    public int $id;
    /** The textual content of the token. */
    public string $text;
    /** The starting line number (1-based) of the token. */
    public int $line;
    /** The starting position (0-based) in the tokenized string. */
    public int $pos;
+   /** Indicates T_WHITESPACE, T_COMMENT, T_DOC_COMMENT, T_OPEN_TAG */
+   public bool $ignorable;
}
```

However, at this point, that's just a nit-pick. In practice I'm not sure how general of an 'optimization' this would be.