@nikic 2020-02-13
Added method implementations to this PR now...

@marcioAlmada So there is a trade-off between memory usage and speed here. You are right that having a separate `ignorable` property is going to be more efficient, because accessing the property is cheaper than calling a method (even if the method implementation here is really cheap). However, having this boolean (1 bit of information...) would add 128 bits of extra memory usage to each token. As memory usage is pretty sensitive here, I think it's not a good tradeoff.