@TysonAndre 2020-02-16
Thanks for working on this - I'm looking forward to the time and memory improvements of this.

I've noticed that if you add instance properties to the subclass of PhpToken, they don't get instantiated to the default - they're uninitialized instead.
I don't know if the class this is based on would have the same issue.

If this is intended, this should be documented in the RFC.
If this is not intended, this should be fixed (e.g. take a slower code path for instantiating subclasses, check EG(exception), or make PhpToken final if that's unexpectedly impractical.)

```php
--TEST--
Extending the PhpToken class with extra properties
--FILE--
<?php

$code = <<<'PHP'
Hello, world!
PHP;

// Observed: When MyPhpToken::getAll() is called, otherProp is absent
// Expected: Properties get initialized to the default constant expressions.
// If default constant expressions are invalid (E.g. MISSING_CONSTANT), it throws.
class MyPhpToken extends PhpToken {
    public $otherProp = 123;
    public function getLoweredText(): string {
        return strtolower($this->text);
    }
}

var_export(MyPhpToken::getAll('hello, world'));
// The below is what actually happens, not the desired behavior
?>
--EXPECT--
array (
  0 => 
  MyPhpToken::__set_state(array(
     'id' => 312,
     'text' => 'hello, world',
     'line' => 1,
     'pos' => 0,
  )),
)
```