@jrfnl 2020-02-16
cc @sebastianbergmann re php-token-stream