@nikic 2020-02-17
@TysonAndre Nice catch! That wasn't intentional, and initialization of extra properties is now fixed. I've taken care to implement this in a way that only adds additional runtime cost if there really are extra properties.